requirements.txt
streamlit
requests
feedparser
python-dotenv
sqlalchemy
agno
groq